refactor: schedule-edit 컴포넌트 폴더 위치 이동
