+++ src/features/schedule-edit/model/format-date-to-string.ts
