+++ src/features/schedule-edit/ui/schedule-edit-repeat.tsx
@@ -8,9 +8,9 @@ import {
   BottomSheetHeader,
   BottomSheetHeaderButton,
   BottomSheetHeaderTitle,
-} from '../bottom-sheet'
-import { Radio, RadioGroup } from '../radio'
-import Text from '../text/text'
+} from '../../../shared/ui/bottom-sheet'
+import { Radio, RadioGroup } from '../../../shared/ui/radio'
+import Text from '../../../shared/ui/text/text'
 
 import ScheduleEditRepeatOption from './schedule-edit-repeat-option'
